changed topic list to table

+++ api/src/main/java/base/api/domain/forum/topics/TopicEntity.java
@@ -10,5 +10,5 @@ public class TopicEntity extends GenericEntity {
 
   private String topicTitle;
   private Integer topicViews;
-  private Long topicCategory;
+  private Integer topicCategory;
 }
